+++ internal/compiler/lookup.rs
@@ -858,7 +858,6 @@ impl<'a> LookupObject for ColorExpression<'a> {
         None.or_else(|| f("brighter", member_function(BuiltinFunction::ColorBrighter)))
             .or_else(|| f("darker", member_function(BuiltinFunction::ColorDarker)))
             .or_else(|| f("transparentize", member_function(BuiltinFunction::ColorTransparentize)))
-            .or_else(|| f("opaque", member_function(BuiltinFunction::ColorOpaque)))
             .or_else(|| f("with-alpha", member_function(BuiltinFunction::ColorWithAlpha)))
             .or_else(|| f("mix", member_function(BuiltinFunction::ColorMix)))
     }